currents are floats

+++ app/db/model.py
@@ -464,8 +464,8 @@ class MEModel(
         "ReconstructionMorphology", foreign_keys=[morphology_id], uselist=False
     )
 
-    holding_current: Mapped[int]
-    threshold_current: Mapped[int]
+    holding_current: Mapped[float]
+    threshold_current: Mapped[float]
 
     emodel_id: Mapped[uuid.UUID] = mapped_column(ForeignKey(f"{EntityType.emodel}.id"))
